+++ src/components/common/Button.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import styled from 'styled-components';
 
 interface ButtonProps {
-  size: 'small' | 'medium' | 'full';
-  color: 'primary';
+  size: 'small' | 'medium' | 'half' | 'full';
+  color: string;
   children: React.ReactNode;
   theme?: {
     colors: {
@@ -26,7 +26,7 @@ const Button = ({ size, color, children, onClick }: ButtonProps) => {
 const StyledButton = styled.button<ButtonProps>`
   padding: ${(props) => getButtonPadding(props.size)};
   width: ${(props) => getButtonWidth(props.size)};
-  background-color: ${(props) => getButtonColor(props)};
+  background-color: ${(props) => getButtonColor(props.color)};
   color: #fff;
   border: none;
   border-radius: 10px;
@@ -36,25 +36,16 @@ const StyledButton = styled.button<ButtonProps>`
   font-weight: bolder;
 
   &:hover {
-    background-color: ${(props) => getButtonHoverColor(props.color)};
+    opacity: 0.8;
   }
 `;
 
-const getButtonColor = (props: ButtonProps) => {
-  switch (props.color) {
-    case 'primary':
-      return '#007DFa';
-    default:
-      return '#007DFa';
-  }
-};
-
-const getButtonHoverColor = (color: string) => {
+const getButtonColor = (color: string) => {
   switch (color) {
     case 'primary':
-      return 'rgba(0, 125, 250, 0.8)';
+      return '#007DFa';
     default:
-      return 'rgba(0, 125, 250, 0.8)';
+      return color;
   }
 };
 
@@ -64,6 +55,8 @@ const getButtonWidth = (size: string) => {
       return '5vw';
     case 'medium':
       return '8vw';
+    case 'half':
+      return '50%';
     case 'full':
       return '100%';
     default:
@@ -77,6 +70,8 @@ const getButtonPadding = (size: string) => {
       return '8px 16px';
     case 'medium':
       return '12px 24px';
+    case 'half':
+      return '14px 27px';
     case 'full':
       return '14px 27px';
     default:
